fix(1-1-restore): fix unit-tests
This fixes unit-tests in 1-1-restore-feature-branch after it was rebased
on master.
